// src/app/modules/integrations/models/ebay-setting.ts
export class EbaySetting{
        applyShippingDiscount: boolean;
        automaticBestOfferAmountRule: any;
        bestOfferAmount: 0;
        bestOfferAmountRule: any;
        combinedInternationalShippingDiscount: boolean;
        combinedShippingDiscount: boolean;
        defaultOption: boolean;
        depth: any;
        dispatchTimeMax: any;
        domesticHandlingTime: any;
        domesticInsurance: any;
        domesticInsuranceValue: any;
        domesticOption: any;
        domesticServiceOne: any;
        domesticServiceOneValue: any;
        domesticServiceOneValueAdditional: any;
        domesticServiceThree: any;
        domesticServiceThreeValue: any;
        domesticServiceThreeValueAdditional: any;
        domesticServiceTwo: any;
        domesticServiceTwoValue: any;
        domesticServiceTwoValueAdditional: any;
        domesticShippingDiscountProfileId: any;
        ebayProductXml: any;
        excludeShipToLocations: any;
        expediteDomesticShipping: boolean;
        freeshipping: boolean;
        getItFast: boolean;
        globalShipping: boolean;
        id: 0;
        intdepth: any;
        internationalFiveShipToLocations: [
          any
        ];
        internationalFourShipToLocations: [
        ];
        internationalInsurance: any;
        internationalInsuranceValue: any;
        internationalOneShipToLocations: [
        ];
        internationalOption: any;
        internationalServiceFive: any;
        internationalServiceFiveValue: any;
        internationalServiceFiveValueAdditional: any;
        internationalServiceFour: any;
        internationalServiceFourValue: any;
        internationalServiceFourValueAdditional: any;
        internationalServiceOne: any;
        internationalServiceOneValue: any;
        internationalServiceOneValueAdditional: any;
        internationalServiceThree: any;
        internationalServiceThreeValue: any;
        internationalServiceThreeValueAdditional: any;
        internationalServiceTwo: any;
        internationalServiceTwoValue: any;
        internationalServiceTwoValueAdditional: any;
        internationalShippingDiscountProfileId: any;
        internationalThreeShipToLocations: [
          any
        ];
        internationalTwoShipToLocations: [
          any
        ];
        intiregularPackage: any;
        intlength: any;
        intpackageType: any;
        intweightHandlingCost: any;
        intweightLbs: any;
        intweightOz: any;
        intwidth: any;
        iregularPackage: any;
        kartzhubUserId: 0;
        length: any;
        locationsFive: any;
        locationsFour: any;
        locationsOne: any;
        locationsThree: any;
        locationsTwo: any;
        name: any;
        packageType: any;
        postageProfileId: any;
        promotionalDiscountProfile: any;
        refundOption: any;
        returnAccepted: boolean;
        returnPolicyDescription: any;
        returnPolicyId: any;
        returnPolicyProfileId: any;
        returnsWithinOption: any;
        shipToLocations: [
          any
        ];
        shipToLocationsList: [
          any
        ];
        shipToRegistrationCountry: boolean;
        shippingCostPaidBy: any;
        shippingCostPaidByOption: any;
        shippingPolicyId: any;
        shiptolocations: any;
        site: any;
        weightHandlingCost: any;
        weightLbs: any;
        weightOz: any;
        width: any
}