// src/app/modules/integrations/models/ebay-policy.ts
export class EbayPolicy{
    amEx: true;
    attributeSetId: 0;
    attributenames: [
      string
    ];
    attributes: [
      
    ];
    attributevalues: [
      string
    ];
    autoPay: true;
    bankDeposit: true;
    businessSeller: true;
    cashOnPickup: true;
    ccaccepted: true;
    cheque: true;
    cod: true;
    creditcard: true;
    defaultOption: true;
    ebayProductXml: string;
    escrow: true;
    id: 0;
    immediatePayment: true;
    kartzhubUserId: 0;
    merchantccamex: true;
    merchantccdiscover: true;
    merchantccvisamastercard: true;
    mocc: true;
    moneyBookers: true;
    moneyOrder: true;
    moneyXferAccepted: true;
    moneyXferAcceptedInCheckout: true;
    moneybookers: true;
    name: string;
    other: true;
    paisaPay: true;
    parisaPayAccepted: true;
    payMate: true;
    payPal: true;
    paymate: true;
    paymentPolicyId: string;
    paymentProfileId: string;
    payonpickup: true;
    paypalChecked: true;
    paypalEmail: string;
    personalCheck: true;
    postalOrder: true;
    postalTransfer: true;
    pricingruleroundvalue: string;
    proPay: true;
    promotionalSaleId: string;
    propay: true;
    refundMethodSelected: string;
    refundOption: string;
    restockingFee: string;
    restockingFeeOption: string;
    restrictedToBusiness: true;
    returnAccepted: true;
    returnDurationSelected: string;
    returnPaidBySelected: string;
    returnPolicyText: string;
    returnsWithinOption: string;
    shippingCostPaidBy: string;
    shippingCostPaidByOption: string;
    showBankAccountNumberInCheckout: true;
    site: string;
    taxState: string;
    taxStates: [
      {
        code: string;
        id: string;
        value: string
      }
    ];
    vatAmount: 0;
    vatpercent: 0

}